parsing update

+++ Include/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: cchabeau <cchabeau@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/06 10:44:02 by cchabeau          #+#    #+#             */
-/*   Updated: 2023/03/30 13:15:53 by cchabeau         ###   ########.fr       */
+/*   Updated: 2023/03/30 14:36:14 by cchabeau         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,6 +17,7 @@
 # include <stdlib.h>
 # include <unistd.h>
 # include <stdarg.h>
+# include <limits.h>
 
 enum e_type
 {
@@ -38,36 +39,36 @@ typedef struct s_stack
 }		t_stack;
 
 //	UTILS
-char	**ft_split(char const *s, char c);
-size_t	ft_strlen(const char *s);
-int		ft_strcmp(const char *s1, const char *s2);
-int		ft_isnumeric(char *str);
-char	*ft_strdup(const char *s);
-size_t	ft_tablen(char **array);
-void	ft_putstr_fd(char *s, int fd);
-int		ft_atoi(const char *str);
+char		**ft_split(char const *s, char c);
+size_t		ft_strlen(const char *s);
+int			ft_strcmp(const char *s1, const char *s2);
+int			ft_isnumeric(char *str);
+char		*ft_strdup(const char *s);
+size_t		ft_tablen(char **array);
+void		ft_putstr_fd(char *s, int fd);
+long int	ft_atoi(const char *str);
 
 //	PRINT
-size_t	ft_convert(char type, va_list ap);
-int		ft_print_char(char c);
-int		ft_printf(const char *format, ...);
-void	ft_print_tab(char **array);
-void	print_stack(t_stack *stack);
+size_t		ft_convert(char type, va_list ap);
+int			ft_print_char(char c);
+int			ft_printf(const char *format, ...);
+void		ft_print_tab(char **array);
+void		print_stack(t_stack *stack);
 
 //	ERROR
-void	free_tab(char **array);
-void	free_stack(t_stack *stack);
-void	ft_error(void *ptr, int type);
+void		free_tab(char **array);
+void		free_stack(t_stack *stack);
+void		ft_error(void *ptr, int type);
 
 //	PARSING
-int		is_double(int nbr, t_stack *stack);
-int		is_validArg(int len, char **array);
-char	**ft_get_arg(int argc, char **argv);
+int			is_double(long int nbr, t_stack *stack);
+int			is_validArg(int len, char **array);
+char		**ft_get_arg(int argc, char **argv);
 
 //	STACK
-t_stack	*init_stack(int	nbr);
-t_stack	*push_stack(int nbr, t_stack *stack);
-t_stack	*pop_stack(t_stack *stack);
-t_stack	*fill_stack(char **array, t_stack *stack);
+t_stack		*init_stack(int	nbr);
+t_stack		*push_stack(int nbr, t_stack *stack);
+t_stack		*pop_stack(t_stack *stack);
+t_stack		*fill_stack(char **array, t_stack *stack);
 
 #endif
